requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.5,<6
djangorestframework-gis>=1.1,<2
django-filter>=25,<26
psycopg[binary]>=3.1,<4
celery[redis]>=5.4,<6
redis>=5.0,<6
